Split controllers in local login/logout and external challenge/callback (#2204)
* split local and external login

* fixed logic bugs around local urls, PKCE check
