+++ src/Host/Views/Account/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@model LoginViewModel
+@model LoginViewModel
 
 <div class="login-page">
     <div class="page-header">
@@ -63,7 +63,8 @@
                             {
                                 <li>
                                     <a class="btn btn-default"
-                                       asp-action="ExternalLogin"
+                                       asp-controller="External"
+                                       asp-action="Challenge"
                                        asp-route-provider="@provider.AuthenticationScheme"
                                        asp-route-returnUrl="@Model.ReturnUrl">
                                         @provider.DisplayName
